requirements.txt
python-dotenv
fastapi
uvicorn[standard]
aiohttp
microsoft-agents-activity
microsoft-agents-hosting-core
microsoft-agents-hosting-aiohttp
microsoft-agents-authentication-msal
